Linting and sql utility location cleanup.
