+++ quasar/sql_utils.py
@@ -6,7 +6,6 @@ import time
 from .database import Database
 from .sa_database import Database as sadb
 from sqlalchemy import create_engine
-from .utils import sql_replace
 from .utils import log
 
 
@@ -60,6 +59,20 @@ class DataFrameDB:
 def run_sql_file_old(file):
     df = DataFrameDB()
     df.run_query(file)
+
+
+def sql_replace(query, datamap):
+    """Used for find/replace variables in sql_run_file function based
+    on a regular pattern.
+    """
+    # Remove any newlines.
+    final_query = query.replace("\n", "")
+    # Based on variables in datamap, replace ':' prepended values
+    # with actual values.
+    for key in datamap:
+        j = ':' + key
+        final_query = final_query.replace(j, datamap[key])
+    return final_query
 
 
 def run_sql_file(file, datamap):
