+++ quasar/utils.py
@@ -7,7 +7,6 @@ logging.getLogger().setLevel(logging.INFO)
 
 # DoSomething Helper Functions - Code Reused Across Lots of our ETL Scripts
 
-
 def strip_str(base_value):
     """Convert value to string and strips special characters.
 
@@ -40,20 +39,6 @@ def log(msg):
 
 def logerr(msg):
     logging.error(msg)
-
-
-def sql_replace(query, datamap):
-    """Used for find/replace variables in sql_run_file function based
-    on a regular pattern.
-    """
-    # Remove any newlines.
-    final_query = query.replace("\n", "")
-    # Based on variables in datamap, replace ':' prepended values
-    # with actual values.
-    for key in datamap:
-        j = ':' + key
-        final_query = final_query.replace(j, datamap[key])
-    return final_query
 
 
 class Duration:
